Moving a few classes, code cleanup
